+++ src/utils/Utils.C
@@ -10,13 +10,26 @@
  **********************************************************************/
 
 #include "Utils.H"
+
+#include <Teuchos_XMLParameterListHelpers.hpp>
+
+#include "EpetraExt_MatrixMatrix.h"
+#include "EpetraExt_HDF5.h"
+
+#include "Epetra_MultiVector.h"
+#include "Epetra_IntVector.h"
+#include "Epetra_Vector.h"
+#include "Epetra_Import.h"
+#include "Epetra_Map.h"
+#include "Epetra_BlockMap.h"
+#include "Epetra_CrsMatrix.h"
+
+#include "TRIOS_Domain.H"
+
 #include "Combined_MultiVec.H"
 #include "ComplexVector.H"
-#include "TRIOS_Domain.H"
-#include "EpetraExt_MatrixMatrix.h"
+
 #include <functional> // for std::hash
-#include <cstdlib>    // for rand();
-#include <Teuchos_XMLParameterListHelpers.hpp>
 
 using ConstIterator = Teuchos::ParameterList::ConstIterator;
 //========================================================================================
